+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - aria-hidden attribute to all Divider components
 
+### Fixed
+
+- Set focus to question heading when question changed
+
 ## [0.2.8] - 2021-12-23
 
 ### Changed